V0.8
- Implement Audio beeping
- Implement Waiting for keypress
- Separate CPU & program into two separate files
