+++ Chip8CSharp/Chip8CSharp.csproj
@@ -4,8 +4,8 @@
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp2.1</TargetFramework>
     <PublishSingleFile>true</PublishSingleFile>
-    <AssemblyVersion>0.7.0.0</AssemblyVersion>
-    <Version>0.7.0.0</Version>
+    <AssemblyVersion>0.8.0.0</AssemblyVersion>
+    <Version>0.8.0.0</Version>
   </PropertyGroup>
 
   <ItemGroup>
